Completely rewrite the API
This completely rewrites the API. The code is now split into multiple modules and
can be installed as a project/library. It uses Flask CLI to provide commands to
create, update and clean database, Flask server to serve the API, and SQLAlchemy
for database. Error handling has also been updated, it now provides support for
a custom logging config and monitoring using Sentry.

Note: The API is incompatible with the previous version and is currently incomplete.

Features available in the rewrite:
* Database management
* List classes, teachers and classrooms
* Get timetable for class, teacher or classroom
* Get empty classrooms
* Get circulars

Features available in the previous version and currently missing in the rewrite:
* Substitution parsing and access

Features missing from both versions:
* Lunch schedule parsing and access
* Lunch and snack menu parsing and access
* Notification system