Completely rewrite the website
This completely rewrites the website. It now uses Vue.js framework with TypeScript
and SASS support. It is more modular, code is split into components and views, and
should be more reliable. It also provides support for Sentry monitoring.

Features available in the rewrite:
* Display timetable
* Display substitutions for classes
* Entity selection and settings

Features available in the previous version and currently missing in the rewrite:
* Better desktop layout
* Pull to refresh
* Force update and last updated date

Features missing from both versions:
* Substitutions for teachers and classrooms
* Lunch schedule access
* Lunch and snack menu access
* Documents access
* Notification system
* Dark theme