Improve Sentry integration for the API
It is now possible to set custom trace sample rates for commands and
requests independently. It is now also possible to store user IPs
on Sentry for requests. Users can opt-out from storing their IPs by
setting Do Not Track or Global Privacy Control headers. It is also
possible to set custom release prefixes and suffixes that can be used
in Sentry organization with multiple projects.