Move checking if document has changed into a separate method
This makes it possible for specific updaters to provide their own login for determining if the document has changed, without downloading the document and checking the hash. This can be useful if there is an easy way to determine if the document has changed, for example by comparing modified dates from the source, or other source-specific logic, so we can skip downloading the document. Even if this method returns that the document has changed, we still download the document and compare hashes.