Add lunch until field for future uses
This will be used in the future to store the lunch until value in the menu updater and display it in the frontend.

This introduces a backward incompatible change. A new nullable time `until` field has been added to the `lunch_menu` table. You need to manually add it or recreate the database.